move wdyr into development
